can be used as es module

+++ bin/cleanAll.js
@@ -7,14 +7,14 @@ const glob_1 = __importDefault(require("glob"));
 const fs_1 = __importDefault(require("fs"));
 const getConfig_1 = __importDefault(require("./getConfig"));
 const setConfig_1 = __importDefault(require("./setConfig"));
+const checkExistFile_1 = __importDefault(require("./checkExistFile"));
 async function cleanAll() {
     const config = await setConfig_1.default(getConfig_1.default());
     const filePath = config.convertedFile;
     glob_1.default(filePath, (error, files) => {
-        if (error) {
-            console.log(error);
-        }
         for (const file of files) {
+            if (!checkExistFile_1.default(file))
+                return;
             fs_1.default.unlinkSync(file);
         }
     });
